reworks vanilla_binary to use usize instead of i32
