--- src/binary_search_01.rs
@@ -1,5 +1,0 @@
-fn vanilla_binary(list: &vec, i: i32) -> i32 {
-    //returns index of target
-    //or -1 if not found
-    0 //test filler value
-}
