+++ src/main.rs
@@ -1,3 +1,5 @@
+pub mod binary01;
+
 fn main() {
     println!("Hello, world!");
 }
